MOAT: Basic features
The MOAT is the Master Of All Things -- 1wire "slave" code for small AVR
processors. Requirements: Timer 0 interrupt, INT0 input to connect 1wire
to, at least 8 MHz clock, 3 kBytes of flash memory and 10 bytes RAM.

This is the initial proof-of-concept check-in. It can read device
information and "console" messages. in other words, this does not do
anything useful. Yet.

The slave code is located at https://github.com/M-o-a-T/owslave.